# add benchmark and use RWLock instead of mutex Lock